requirements.txt
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
cssselect==1.3.0
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2.6
django-cache-url==3.4.5
djangorestframework==3.16.1
environs==14.3.0
executing==2.2.1
filelock==3.19.1
greenlet==3.2.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
lxml==6.0.2
marshmallow==4.0.1
matplotlib-inline==0.1.7
packaging==25.0
parsel==1.10.0
parso==0.8.5
playwright==1.55.0
prompt_toolkit==3.0.52
Protego==0.5.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
python-dotenv==1.1.1
queuelib==1.8.0
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
scrapy-playwright==0.0.44
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
stack-data==0.6.3
tldextract==5.3.0
traitlets==5.14.3
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
w3lib==2.3.1
wcwidth==0.2.14
zope.interface==8.0.1
